// desktop/src/types.ts
type TMaybe<T> = T | null | undefined
// ...
export type TProviderID = string
export type TOptionID = string
export type TWithProviderID = Readonly<{ providerID: TProviderID }>
export type TProviders = Record<TProviderID, TProvider>
export type TProvider = Readonly<{
  config: TMaybe<TProviderConfig>
  default: TMaybe<boolean>
  state: TMaybe<
    Readonly<{
      initialized: TMaybe<boolean>
      singleMachine: TMaybe<boolean>
      creationTimestamp: TMaybe<string>
      options: TMaybe<TProviderOptions>
    }>
  >
}> & {
  isProxyProvider: boolean
}
export type TNamedProvider = TProvider & Readonly<{ name: string }>
export type TProviderConfig = Readonly<{
  name: TMaybe<string>
  version: TMaybe<string>
  source: TMaybe<TProviderSource>
  description: TMaybe<string>
  optionGroups: TProviderOptionGroup[]
  options: TProviderOptions
  icon: TMaybe<string>
  home: TMaybe<string>
  exec:
    | TMaybe<Record<string, readonly string[]> & { proxy: never; daemon: never }>
    | TMaybe<{ proxy: TMaybe<Record<string, readonly string[]>>; daemon: never }>
    | TMaybe<{ daemon: TMaybe<Record<string, readonly string[]>>; proxy: never }>
}>
export type TProviderOptionGroup = Readonly<{
  name: TMaybe<string>
  options: TMaybe<string[]>
  defaultVisible: TMaybe<boolean>
}>
export type TProviderSource = Readonly<{
  internal: TMaybe<boolean>
  github: TMaybe<string>
  file: TMaybe<string>
  url: TMaybe<string>
  raw: TMaybe<string>
}>
export type TProviderOptions = Record<string, TProviderOption>
export type TProviderOption = Readonly<{
  // Value is the options current value
  value: TMaybe<string>
  // Children are the children generated by this option
  children: TMaybe<string[]>
  // If value is a password
  password: TMaybe<boolean>
  // DisplayName of the option, preferred over the option name by a supporting tool.
  displayName: TMaybe<string>
  // A description of the option displayed to the user by a supporting tool.
  description: TMaybe<string>
  // If required is true and the user doesn't supply a value, devsy will ask the user
  required: TMaybe<boolean>
  // Allowed values for this option.
  enum: TMaybe<TOptionEnum[]>
  // Suggestions are suggestions to show in the Devsy UI for this option
  suggestions: TMaybe<string[]>
  // Hidden specifies if the option should be hidden
  hidden: TMaybe<boolean>
  // Local means the variable is not resolved immediately and instead later when the workspace / machine was created.
  local: TMaybe<boolean>
  // Default value if the user omits this option from their configuration.
  default: TMaybe<string>
  // Command is the command to run to specify an option
  command: TMaybe<string>
  // SubOptionsCommand is the command to retrieve sub options
  subOptionsCommand: TMaybe<string>
  // Type is the provider option type. Can be one of: string, multiline, duration, number or boolean. Defaults to string
  type: TMaybe<"string" | "duration" | "number" | "boolean" | "multiline">
  // Mutable specifies if an option can be changed on the workspace or machine after creating it
  mutable: TMaybe<boolean>
}>
export type TOptionEnum = Readonly<{
  value: TMaybe<string>
  displayName: TMaybe<string>
}>

export type TAddProviderConfig = Readonly<{
  name?: TProviderConfig["name"]
}>
// ...
export type TCheckProviderUpdateResult = Readonly<{
  updateAvailable: boolean
  latestVersion?: string
}>